+++ src/main/java/org/grnet/pidmr/exception/ModeIsNotSupported.java
@@ -1,0 +1,8 @@
+package org.grnet.pidmr.exception;
+
+public class ModeIsNotSupported extends RuntimeException{
+
+    public ModeIsNotSupported(String message) {
+        super(message);
+    }
+}